critical fix in convergence time of recurrences mapper (#174)
We **desperately** need rigorous tests for the convergence time estimation of `AttractorsViaRecurrences`. The fact that I've "fixed" this function 3 times, yet was still giving totally wrong results is a testament to that. I've been sloppy to not write tests and someone (me or someone else) needs to contribute rigorous tests for this function and this mapper.